pyproject.toml
[tool.poetry]
name = "scarabbot"
version = "0.1.0"
description = ""
authors = ["Soulsender <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
requests = "^2.25.1"
Markdown = "^3.3.4"
repltalk = "^1.0.0"
py-cord = "^1.7.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"